Fix Negative Cargo Bug
I believe this was the cause of negative material in a ships cargo.
Basically the if statement was incorrect and it was going to the else
which would distribute the remaining space into the amounts to buy from
each player even though it was'nt time to distribute it. Now it should
properly execute until it gets to the point that the remaining space
needs to be distributed.
closes #207